requirements.txt
asttokens==2.4.1
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.50.0
imageio==2.34.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.3
matplotlib-inline==0.1.6
mplcursors==0.5.3
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
Pygments==2.17.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
scikit-image==0.22.0
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
tifffile==2024.2.12
tornado==6.4
traitlets==5.14.2
tzdata==2024.1
wcwidth==0.2.13
